separate settings card widget
