+++ lib/ui/widgets/account/account_widget.dart
@@ -86,44 +86,6 @@ class _BodyPersonalWidgetState extends State<BodyPersonalWidget> {
   }
 }
 
-class SettingsCardWidget extends StatefulWidget {
-  const SettingsCardWidget({Key? key}) : super(key: key);
-
-  @override
-  State<SettingsCardWidget> createState() => _SettingsCardWidgetState();
-}
-
-class _SettingsCardWidgetState extends State<SettingsCardWidget> {
-  // bool themeColor = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return Consumer<ModelTheme>(
-      builder: (context, ModelTheme themeNotifier, child) {
-        return SwitchListTile(
-          activeColor: Colors.pinkAccent,
-          inactiveThumbColor: Colors.grey,
-          value: themeNotifier.isDark,
-          onChanged: (bool value) {
-            setState(
-              () {
-                themeNotifier.isDark = value;
-              },
-            );
-          },
-          title: const Text(
-            'Change color theme',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
 class NotificationsCardWidget extends StatefulWidget {
   const NotificationsCardWidget({Key? key}) : super(key: key);
